+++ rana_qgis_plugin/widgets/rana_browser.py
@@ -34,6 +34,7 @@ from qgis.PyQt.QtGui import (
 )
 from qgis.PyQt.QtSvg import QSvgWidget
 from qgis.PyQt.QtWidgets import (
+    QComboBox,
     QDialog,
     QDialogButtonBox,
     QFileDialog,
@@ -77,6 +78,7 @@ from rana_qgis_plugin.simulation.threedi_calls import (
 from rana_qgis_plugin.utils import (
     NumericItem,
     convert_to_local_time,
+    convert_to_timestamp,
     display_bytes,
     elide_text,
     get_file_icon_name,
@@ -85,12 +87,14 @@ from rana_qgis_plugin.utils import (
 )
 from rana_qgis_plugin.utils_api import (
     get_tenant_file_descriptor,
+    get_tenant_id,
     get_tenant_project_file,
     get_tenant_project_file_history,
     get_tenant_project_files,
     get_tenant_projects,
     get_threedi_schematisation,
     get_user_image,
+    get_user_info,
 )
 from rana_qgis_plugin.utils_settings import base_url
 from rana_qgis_plugin.utils_spatial import get_bbox_area_in_m2
@@ -100,6 +104,8 @@ from rana_qgis_plugin.widgets.utils_file_action import (
     get_file_actions_for_data_type,
 )
 from rana_qgis_plugin.widgets.utils_icons import (
+    AvatarCache,
+    ContributorAvatarsDelegate,
     create_user_image,
     get_icon_from_theme,
     get_icon_label,
@@ -562,7 +568,8 @@ class FileView(QWidget):
         rows.append([file_icon_label, self.filename_edit, QLabel(size_str)])
         # line 2: user icon - user name - commit msg - time
         # This is broken (or the stuff above)
-        user_image = get_user_image(self.communication, selected_file)
+        # TODO: use avatar cache or whater should be used
+        user_image = get_user_image(self.communication, selected_file["user"]["id"])
         if user_image:
             user_icon_label = get_icon_label(create_user_image(user_image))
         else:
@@ -1020,18 +1027,26 @@ class ProjectsBrowser(QWidget):
     busy = pyqtSignal()
     ready = pyqtSignal()
 
-    def __init__(self, communication, parent=None):
+    def __init__(self, communication, avatar_cache, parent=None):
         super().__init__(parent)
         self.communication = communication
         self.projects = []
+        self.users = []
         self.filtered_projects = []
         self.current_page = 1
         self.items_per_page = 100
         self.project = None
-        self.setup_ui()
+        self.avatar_cache = avatar_cache
+        # collect data
         self.fetch_projects()
+        self.update_users()
+        # start thread to retrieve actual avatars
+        self.avatar_cache.avatar_changed.connect(self.update_avatar)
+        self.avatar_cache.update_users_in_thread(self.users)
+        self.setup_ui()
+        self.populate_contributors()
         self.populate_projects()
-        self.projects_tv.header().setSortIndicator(1, Qt.SortOrder.DescendingOrder)
+        self.projects_tv.header().setSortIndicator(2, Qt.SortOrder.AscendingOrder)
 
     def set_project_from_id(self, project_id: str):
         for project in self.projects:
@@ -1044,13 +1059,37 @@ class ProjectsBrowser(QWidget):
         self.projects_search = QLineEdit()
         self.projects_search.setPlaceholderText("🔍 Search for project by name")
         self.projects_search.textChanged.connect(self.filter_projects)
+        self.projects_search.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Fixed)
+        # Create filter by contributor box
+        self.contributor_filter = QComboBox()
+        self.contributor_filter.setSizePolicy(QSizePolicy.Expanding, QSizePolicy.Fixed)
+        self.contributor_filter.setEditable(True)
+        self.contributor_filter.setInsertPolicy(QComboBox.InsertPolicy.NoInsert)
+        # add auto complete
+        self.contributor_filter.completer().setCaseSensitivity(
+            Qt.CaseSensitivity.CaseInsensitive
+        )
+        # setup placeholder and reset filter on no user selected
+        if self.contributor_filter.lineEdit():
+            self.contributor_filter.lineEdit().setPlaceholderText("All contributors")
+            self.contributor_filter.lineEdit().textChanged.connect(
+                self._on_contributor_filter_text_changed
+            )
+        self.contributor_filter.currentIndexChanged.connect(self.filter_projects)
         # Create tree view with project files and model
         self.projects_model = QStandardItemModel()
         self.projects_tv = QTreeView()
         self.projects_tv.setModel(self.projects_model)
         self.projects_tv.setSortingEnabled(True)
         self.projects_tv.header().setSortIndicatorShown(True)
-        self.projects_tv.header().setSortIndicator(1, Qt.SortOrder.AscendingOrder)
+        self.projects_tv.setContextMenuPolicy(Qt.ContextMenuPolicy.CustomContextMenu)
+        self.projects_tv.customContextMenuRequested.connect(self.show_context_menu)
+        self.projects_tv.header().sortIndicatorChanged.connect(self.sort_projects)
+        self.projects_model.setHorizontalHeaderLabels(
+            ["Project Name", "Contributors", "Last activity"]
+        )
+        avatar_delegate = ContributorAvatarsDelegate(self.projects_tv)
+        self.projects_tv.setItemDelegateForColumn(1, avatar_delegate)
         # self.projects_tv.clicked.connect(self.select_project)
         self.projects_tv.doubleClicked.connect(self.select_project)
         # Create navigation buttons
@@ -1062,6 +1101,7 @@ class ProjectsBrowser(QWidget):
         # Organize widgets in layouts
         top_layout = QHBoxLayout()
         top_layout.addWidget(self.projects_search)
+        top_layout.addWidget(self.contributor_filter)
         pagination_layout = QHBoxLayout()
         pagination_layout.addWidget(self.btn_previous)
         pagination_layout.addWidget(
@@ -1074,62 +1114,240 @@ class ProjectsBrowser(QWidget):
         layout.addLayout(pagination_layout)
         self.setLayout(layout)
 
+    def _on_contributor_filter_text_changed(self, text):
+        # reset contribute filter with empty text
+        if not text:
+            self.contributor_filter.setCurrentIndex(-1)
+
     def fetch_projects(self):
         self.projects = get_tenant_projects(self.communication)
+
+    def update_users(self):
+        self.users = list(
+            {
+                contributor["id"]: contributor
+                for project in self.projects
+                for contributor in project["contributors"]
+            }.values()
+        )
 
     def refresh(self):
         self.current_page = 1
         self.fetch_projects()
-        search_text = self.projects_search.text()
-        if search_text:
-            self.filter_projects(search_text, clear=True)
-            return
-        self.populate_projects(clear=True)
-        self.projects_tv.header().setSortIndicator(1, Qt.SortOrder.DescendingOrder)
+        self.update_users()
+        self.avatar_cache.update_users_in_thread(self.users)
+        if self.filter_active:
+            self.filter_projects()
+        else:
+            self.populate_projects()
+        self.populate_contributors()
+        self.projects_tv.header().setSortIndicator(2, Qt.SortOrder.DescendingOrder)
         self.projects_refreshed.emit()
 
-    def filter_projects(self, text: str, clear: bool = False):
-        self.current_page = 1
-        if text:
+    @property
+    def filter_active(self):
+        return (
+            self.projects_search.text() or self.contributor_filter.currentIndex() >= 0
+        )
+
+    def filter_projects(self):
+        if not self.filter_active:
+            self.filtered_projects = self.projects
+        else:
+            # create all filters
+            project_filters = [
+                self.get_projects_filtered_by_name,
+                self.get_projects_filtered_by_contributor,
+            ]
+            # collect all project ids that are included in each active filter
+            project_ids = [
+                {project["id"] for project in filter_func()}
+                for filter_func in project_filters
+            ]
+            # Find project ids that are included in all filters
+            common_ids = set.intersection(*project_ids)
             self.filtered_projects = [
+                project for project in self.projects if project["id"] in common_ids
+            ]
+        self.populate_projects()
+
+    def get_projects_filtered_by_name(self):
+        text = self.projects_search.text()
+        if text:
+            return [
                 project
                 for project in self.projects
                 if text.lower() in project["name"].lower()
             ]
         else:
-            self.filtered_projects = []
-        self.populate_projects(clear=clear)
+            return self.projects
 
-    @staticmethod
-    def _process_project_item(project: dict) -> list[QStandardItem, NumericItem]:
+    def get_projects_filtered_by_contributor(self):
+        selected_user_id = self.contributor_filter.currentData()
+        if selected_user_id is None:
+            return self.projects
+        else:
+            selected_projects = []
+            for project in self.projects:
+                contributors = [
+                    contributor["id"] for contributor in project.get("contributors", [])
+                ]
+                if selected_user_id in contributors:
+                    selected_projects.append(project)
+            return selected_projects
+
+    def sort_projects(self, column_index: int, order: Qt.SortOrder):
+        self.current_page = 1
+        key_funcs = [
+            lambda project: project["name"].lower(),
+            None,
+            lambda project: -convert_to_timestamp(project["last_activity"]),
+        ]
+        key_func = key_funcs[column_index]
+        if key_func:
+            self.projects.sort(
+                key=key_func, reverse=(order == Qt.SortOrder.DescendingOrder)
+            )
+        if self.filter_active:
+            self.filter_projects()
+            return
+        self.populate_projects()
+
+    def show_context_menu(self, position):
+        # Get the index under the cursor
+        index = self.projects_tv.indexAt(position)
+
+        # Check if we clicked on a valid item and it's in column 0
+        if not index.isValid() or index.column() != 0:
+            return
+
+        # Get the project data
+        project_item = self.projects_model.itemFromIndex(index)
+        project = project_item.data(Qt.ItemDataRole.UserRole)
+
+        # Create context menu
+        menu = QMenu(self)
+
+        # Add menu actions
+        open_in_qgis = menu.addAction("Open project in QGIS")
+        open_in_web = menu.addAction("Open project in Rana Web")
+
+        # Show the menu and get the selected action
+        action = menu.exec(self.projects_tv.viewport().mapToGlobal(position))
+
+        # Handle the selected action
+        if action == open_in_qgis:
+            self.select_project(index)
+        elif action == open_in_web:
+            link = f"{base_url()}/{get_tenant_id()}/projects/{project['code']}"
+            QDesktopServices.openUrl(QUrl(link))
+
+    def process_project_item(
+        self, project: dict
+    ) -> list[QStandardItem, QStandardItem, NumericItem]:
         project_name = project["name"]
+        # Process project name into item
         name_item = QStandardItem(project_name)
         name_item.setToolTip(project["code"])
         name_item.setData(project, role=Qt.ItemDataRole.UserRole)
+        # Process last activity time into item
         last_activity_item = get_timestamp_as_numeric_item(project["last_activity"])
-        return [name_item, last_activity_item]
+        # process list of contributors into items
+        contributors_item = QStandardItem()
+        contributors_data = []
+        for i, contributor in enumerate(project.get("contributors", [])):
+            avatar = (
+                self.avatar_cache.get_avatar_for_user(contributor) if i < 3 else None
+            )
+            contributors_data.append(
+                {
+                    "id": contributor["id"],
+                    "name": contributor["given_name"]
+                    + " "
+                    + contributor["family_name"],
+                    "avatar": avatar,
+                }
+            )
+        contributors_item.setData(contributors_data, Qt.ItemDataRole.UserRole)
+        contributors_item.setData(-1, Qt.ItemDataRole.InitialSortOrderRole)
+        return [name_item, contributors_item, last_activity_item]
 
-    def populate_projects(self, clear: bool = False):
-        if clear:
-            self.projects_model.clear()
-        self.projects_model.removeRows(0, self.projects_model.rowCount())
-        header = ["Project Name", "Last activity"]
-        self.projects_model.setHorizontalHeaderLabels(header)
-
+    def populate_projects(self):
         # Paginate projects
-        search_text = self.projects_search.text()
-        projects = self.filtered_projects if search_text else self.projects
+        projects = self.filtered_projects if self.filter_active else self.projects
         start_index = (self.current_page - 1) * self.items_per_page
         end_index = start_index + self.items_per_page
         paginated_projects = projects[start_index:end_index]
-
-        # Add paginated projects to the project model
-        for project in paginated_projects:
-            self.projects_model.appendRow(self._process_project_item(project))
-        for i in range(len(header)):
+        # Prepare data for adding
+        processed_rows = [
+            self.process_project_item(project) for project in paginated_projects
+        ]
+        # Remove current data just in time
+        self.projects_model.removeRows(0, self.projects_model.rowCount())
+        # Populate model with new data
+        for row in processed_rows:
+            self.projects_model.invisibleRootItem().appendRow(row)
+        for i in range(self.projects_tv.header().count()):
             self.projects_tv.resizeColumnToContents(i)
         self.projects_tv.setColumnWidth(0, 300)
         self.update_pagination(projects)
+
+    def update_avatar(self, user_id: str):
+        avatar = self.avatar_cache.get_avatar_from_cache(user_id)
+        # Update contributor_filter
+        index = self.contributor_filter.findData(user_id)
+        if index != -1:  # -1 means not found
+            self.contributor_filter.setItemIcon(index, QIcon(avatar))
+        # Update projects model
+        root = self.projects_model.invisibleRootItem()
+        for row in range(root.rowCount()):
+            contributors_item = root.child(row, 1)
+            contributors_data = contributors_item.data(Qt.ItemDataRole.UserRole)
+            # Check if any contributors in the data match the updated one
+            match = next(
+                (
+                    contributor
+                    for contributor in contributors_data
+                    if contributor["id"] == user_id
+                ),
+                None,
+            )
+            if match and match["avatar"]:
+                match["avatar"] = avatar
+                contributors_item.setData(contributors_data, Qt.ItemDataRole.UserRole)
+
+    def populate_contributors(self):
+        # Collect all unique contributors to the projects
+        all_contributors = {
+            contributor["id"]: contributor
+            for project in self.projects
+            for contributor in project["contributors"]
+        }
+        # Sort users by name, starting with the current user
+        my_info = get_user_info(self.communication)
+        if my_info and my_info.get("sub") in all_contributors:
+            my_id = my_info["sub"]
+            my_user = [all_contributors.pop(my_id)]
+        else:
+            my_id = None
+            my_user = []
+        sorted_users = my_user + sorted(
+            all_contributors.values(),
+            key=lambda x: f"{x['given_name']} {x['family_name']}".lower(),
+        )
+        # Update combo box items
+        self.contributor_filter.blockSignals(True)
+        self.contributor_filter.clear()
+        for user in sorted_users:
+            display_name = f"{user['given_name']} {user['family_name']}"
+            if user["id"] == my_id:
+                display_name += " (You)"
+            user_image = self.avatar_cache.get_avatar_for_user(user)
+            self.contributor_filter.addItem(
+                QIcon(user_image), display_name, userData=user["id"]
+            )
+        self.contributor_filter.setCurrentIndex(-1)
+        self.contributor_filter.blockSignals(False)
 
     def update_pagination(self, projects: list):
         total_items = len(projects)
@@ -1360,6 +1578,7 @@ class RanaBrowser(QWidget):
         super().__init__()
         self.last_refresh_time = time.time()
         self.communication = communication
+        self.avatar_cache = AvatarCache(communication)
         self.setup_ui()
         self.refresh_timer = QTimer()
         self.refresh_timer.timeout.connect(self.auto_refresh)
@@ -1405,7 +1624,7 @@ class RanaBrowser(QWidget):
         banner.setFixedHeight(height)
         self.logo_label = banner
         self.logo_label.installEventFilter(self)
-
+        self.window().installEventFilter(self)
         top_layout.addWidget(self.breadcrumbs, 0, 0, 1, 3)
         spacer = QSpacerItem(40, 20, QSizePolicy.Expanding, QSizePolicy.Minimum)
         top_layout.addItem(spacer, 0, 3, 1, 1)
@@ -1425,7 +1644,9 @@ class RanaBrowser(QWidget):
         # Setup widgets that populate the rana widget
         file_signals = FileActionSignals()
         self.projects_browser = ProjectsBrowser(
-            communication=self.communication, parent=self
+            communication=self.communication,
+            avatar_cache=self.avatar_cache,
+            parent=self,
         )
         self.files_browser = FilesBrowser(
             communication=self.communication, file_signals=file_signals, parent=self
@@ -1590,7 +1811,7 @@ class RanaBrowser(QWidget):
         )
 
     def eventFilter(self, obj, event):
-        if obj == self.logo_label and event.type() == QEvent.MouseButtonPress:
+        if event.type() == QEvent.MouseButtonPress and obj == self.logo_label:
             link = base_url()
             QDesktopServices.openUrl(QUrl(link))
         elif event.type() == QEvent.WindowActivate: